Confocal transfers data through serialization and deserialization, converting the data from object to linear byte stream, which is faster and more convenient.
The ConfocalHistoryEntry class will check the confocal parameter and check if the created image matches each other. It has the option to store data to confocal history and restore data.

The default position is at the center. Position change can only be executed when scanner is not running.

When saving the data, the code will create figures according to whether it's xy scan or depth scan, and load the data into the saved file.

Signed-off-by: Andrew Tong <ytong37@wisc.edu>