commented on the odmrlogic file
Use frequency sweep or list mode during odmr scan.
The input from the user will let the module check if it's running or not, and make the corresponding changes. In the fit logic, this odmr logic file uses gaussian and lorentzian fitting, designated for modes distribution systems and single mode systems respectively. The codes contain setting the trigger polarity of microwave trigger for list and sweep modes.

When starting the module, it will set the default logic variables to initial values, and gets limits from the hardware and set data arrays for recording data. When "clear" is given to the module, it erases the previous plot and start from a new time afterwards. When the module is deactivated, it will turno off microwave source and disconnect signals from hardware.

Signed-off-by: Andrew Tong <ytong37@wisc.edu>